app/functions/requirements.txt
azure-functions==1.17.0
azure-storage-queue==12.6.0
pandas==2.1.0
